+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 See Plugin [README](https://github.com/NotePlan/plugins/blob/main/codedungeon.Toolbox/README.md) for details on available commands and use case.
 
+## [3.11.0] - 2022-07-24 (@codedungeon)
+
+- Added `logInfo`, `logError`, `logWarn`, `logDebug` helpers, support plugin `Debugging` section (@dwertheimer)
+- Updated plugin starter (`npc plugin:create`) to include `Debugging` section in settings (@codedungeon)
+
 ## [3.10.2] - 2022-07-10 (@codedungeon)
 
 - added `np.Templating/lib/support/ejs.js` to `.eslintrc :: ignorePatterns`